// web/src/gen/model/pipelinePackageDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { IReadOnlyListOfPipelinePackageItemDto } from "./iReadOnlyListOfPipelinePackageItemDto";
import type { PipelinePackageDependenciesDto } from "./pipelinePackageDependenciesDto";
import type { PipelinePackageMetadataDto } from "./pipelinePackageMetadataDto";

export interface PipelinePackageDto {
  formatVersion: string;
  bundleType: string;
  exportedAt: string;
  metadata: PipelinePackageMetadataDto;
  pipelines: IReadOnlyListOfPipelinePackageItemDto;
  dependencies: PipelinePackageDependenciesDto;
}