// web/src/gen/model/validatePipelinePackageResponseDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation Studio
 * Digital Content Creation Automation Studio Engine & API
 * OpenAPI spec version: 1.0.0
 */
import type { IReadOnlyListOfPipelineContentTypeDependencyPreviewDto } from "./iReadOnlyListOfPipelineContentTypeDependencyPreviewDto";
import type { IReadOnlyListOfPipelinePackageItemPreviewDto } from "./iReadOnlyListOfPipelinePackageItemPreviewDto";
import type { IReadOnlyListOfPipelineScriptDependencyPreviewDto } from "./iReadOnlyListOfPipelineScriptDependencyPreviewDto";
import type { IReadOnlyListOfString } from "./iReadOnlyListOfString";
import type { PipelinePackageMetadataDto } from "./pipelinePackageMetadataDto";

export interface ValidatePipelinePackageResponseDto {
  isValid: boolean;
  validationErrors: IReadOnlyListOfString;
  packageMetadata: PipelinePackageMetadataDto;
  pipelines: IReadOnlyListOfPipelinePackageItemPreviewDto;
  scripts: IReadOnlyListOfPipelineScriptDependencyPreviewDto;
  contentTypes: IReadOnlyListOfPipelineContentTypeDependencyPreviewDto;
}