// web/src/gen/model/importPipelinePackageResponseDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation Studio
 * Digital Content Creation Automation Studio Engine & API
 * OpenAPI spec version: 1.0.0
 */
import type { IReadOnlyListOfImportedPipelineSummaryDto } from "./iReadOnlyListOfImportedPipelineSummaryDto";

export interface ImportPipelinePackageResponseDto {
  importedPipelinesCount: number;
  pipelines: IReadOnlyListOfImportedPipelineSummaryDto;
  installedScriptsCount: number;
}