// web/src/gen/model/pipelinePackageItemPreviewDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { IReadOnlyListOfString } from "./iReadOnlyListOfString";

export interface PipelinePackageItemPreviewDto {
  bundleId: string;
  name: string;
  isRoot: boolean;
  importOrder: number;
  dependsOn: IReadOnlyListOfString;
  nodeCount: number;
  edgeCount: number;
  hasNameConflict: boolean;
  suggestedName: string;
}