// web/src/gen/model/pipelineScriptDependencyPreviewDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation Studio
 * Digital Content Creation Automation Studio Engine & API
 * OpenAPI spec version: 1.0.0
 */

export interface PipelineScriptDependencyPreviewDto {
  key: string;
  name: string;
  executor: string;
  contentHash: string;
  fileName: string;
  alreadyExists: boolean;
}